requirement.txt
asgiref==3.10.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==5.2.8
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
fonttools==4.60.1
iniconfig==2.3.0
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
mpmath==1.3.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tzdata==2025.2
